+++ plugins/woocommerce-admin/tests/e2e/specs/activate-and-setup/complete-onboarding-wizard.test.tsx
@@ -1,7 +1,9 @@
 const {
 	testAdminOnboardingWizard,
 	testSelectiveBundleWCPay,
+	testDifferentStoreCurrenciesWCPay,
 } = require( '@woocommerce/admin-e2e-tests' );
 
 testAdminOnboardingWizard();
 testSelectiveBundleWCPay();
+testDifferentStoreCurrenciesWCPay();